Changed the camera default orientation and added a GameObject
This object is supposed to mimic a spoon and will be used for testing the effectivity of the giroscope aswell as Unity's physics engine.
Next: will write a script to test the physics by making the "spoon" rotate and add another GameObject like a sphere to check how it works.