Fixed slow faling.
Fixed slow falling based on an aswer from the Unity community:
https://answers.unity.com/questions/1385350/my-objects-are-falling-to-slowly-when-a-rigidbody.html

Turns out the problem was the scaling. I was making a 1 meter wide egg
fall, but I was looking at it from a distance that made it look like a
5cm egg, and that makes it look like it's falling really slowly, even in
real life (tho in real life there are other clues to determine the size
of the object such as the "material").