Attempt to prevent the spoon to be affected by...
the z axis rotation (I think).
Currently, the spoon will not rotate with that axis, but all the other
rotations will be messed up.
I am not sure what to do. Here I tried following something from the
Internet but I'm not so sure it works.
Maybe theres a way to calculate something between the z and the other
axises. Need to try to do that.