Probably found a fix for the problem on #2.
By applying that's shown here:
https://forum.unity.com/threads/gyro-rotation-but-in-local-space.989537/#post-6428510
I was able to rotate the object sucessfully ignoring the other axises.

One problem remains however: I have no clue about how the code
accomplishes this and the object spawns in a weird rotation. This means
I need to understand it and proceed to make the necessary changes.